+++ src/RemoteFileTreeProvider.ts
@@ -18,14 +18,27 @@ export class RemoteFileTreeProvider implements TreeDataProvider<TreeItem> {
     constructor(config: vscode.WorkspaceConfiguration) {
     }
 
+
+
     public connect(config: vscode.WorkspaceConfiguration, connectConfig: ConnConfig) {
+        var self = this;
+
+        var onConnect = function() {
+            if(connectConfig.defaultPath) {
+                self.changePath(connectConfig.defaultPath);
+            }
+        }
+
         this.config = config;
-        this.remoteConnection = new RemoteConnection(this.config, connectConfig, this._onDidChangeTreeData);
+        this.remoteConnection = new RemoteConnection(this.config, connectConfig, this._onDidChangeTreeData, onConnect);
+       // var x = this.onDidChangeTreeData(onConnect);
     }
+
 
     public endSession() {
         return this.remoteConnection.end().then((res) => {
             this.remoteConnection.connStatus = ConnectionStatus.Disconnected;
+            this.remoteConnection.statusBar.dispose();
             this._onDidChangeTreeData.fire();
         });
     }